backend/requirements.txt
Django==4.2.9
django-filter==23.4
django-debug-toolbar==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-spectacular==0.27.1
pre-commit==3.6.2
psycopg2-binary==2.9.9
pytest==7.4.4
pytest-django==4.7.0
django-cors-headers==4.3.1
channels[daphne]==4.0.0
channels-redis==4.2.0

# redis==5.0.1
# celery==5.3.6
# django-celery-beat==2.5.0
